pyproject.toml
[tool.poetry]
name = "browserbase"
version = "0.3.0"
description = "Browserbase Python SDK"
authors = ["Browserbase <support@browserbase.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1"
pydantic = ">=2.7.1"
httpx = ">=0.27.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
playwright = "^1.46.0"
selenium = ">=4.23.0"
black = "^24.8.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project.urls]
Homepage = "https://browserbase.com"
Source = "https://github.com/browserbase/python-sdk"
Issues = "https://github.com/browserbase/python-sdk/issues"

[virtualenvs]
create = true
in-project = true